+++ org.servicifi.gelato.analysis.framework/src-gen/org/servicifi/gelato/analysis/framework/analyses/AnalysisResult.java
@@ -30,12 +30,12 @@ public interface AnalysisResult extends EObject, Comparable {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Label</em>' attribute.
-	 * @see #setLabel(long)
+	 * @see #setLabel(double)
 	 * @see org.servicifi.gelato.analysis.framework.analyses.AnalysesPackage#getAnalysisResult_Label()
 	 * @model required="true"
 	 * @generated
 	 */
-	long getLabel();
+	double getLabel();
 
 	/**
 	 * Sets the value of the '{@link org.servicifi.gelato.analysis.framework.analyses.AnalysisResult#getLabel <em>Label</em>}' attribute.
@@ -45,6 +45,6 @@ public interface AnalysisResult extends EObject, Comparable {
 	 * @see #getLabel()
 	 * @generated
 	 */
-	void setLabel(long value);
+	void setLabel(double value);
 
 } // AnalysisResult
